+++ apps/hyperlocalise-web/src/lib/providers/organization-external-tms-provider-credentials.ts
@@ -15,10 +15,8 @@ import {
 } from "@/lib/providers/tms-capabilities";
 import { listProviderWebhookSubscriptionSummaries } from "@/lib/providers/provider-webhook-subscription-manager";
 import type { ProviderWebhookSubscriptionSummary } from "@/lib/providers/provider-webhook-subscription-types";
-import {
-  assertProviderUrlResolvable,
-  normalizeProviderBaseUrl,
-} from "@/lib/providers/provider-url-safety";
+import { assertProviderUrlResolvable } from "@/lib/providers/provider-url-resolve";
+import { normalizeProviderBaseUrl } from "@/lib/providers/provider-url-safety";
 import { resolvePhraseBaseUrl } from "@/lib/providers/phrase/phrase-base-url";
 
 export type ExternalTmsProviderKind = "crowdin" | "smartling" | "phrase" | "lokalise";
