+++ apps/hyperlocalise-web/src/lib/providers/provider-url-resolve.ts
@@ -1,0 +1,18 @@
+import { normalizeHostname } from "@/lib/security/ssrf-guard";
+import { assertResolvablePublicHost } from "@/lib/security/ssrf-guard-dns";
+
+import { isSafeProviderUrl } from "./provider-url-safety";
+
+export async function assertProviderUrlResolvable(url: string): Promise<void> {
+  const parsed = new URL(url);
+  if (!isSafeProviderUrl(parsed)) {
+    throw new Error("Provider URL is invalid or unsafe.");
+  }
+
+  const hostname = normalizeHostname(parsed.hostname);
+  if (hostname.endsWith(".test")) {
+    return;
+  }
+
+  await assertResolvablePublicHost(hostname);
+}
